Mister Paver project
This is a project for a landscaping company in Vancouver. In this time I had the opportunity to try new things, like using an image slider, making a gallery page using a grid container, I got involved with more HTML attributes.

I worked on mantain a cleaner code, more efficient, and to keep a global structure to apply names to the html elements to code the least possible and to try to have a format in all the pages through one css file.